data: Move line-id file persistence
Line-selection parsing is a core concern, but processed line IDs are
persisted workflow state used by include, skip, and hunk refresh paths.

Keeping read/write helpers in core.line_selection made core import file I/O
and blurred neutral selection parsing with session bookkeeping.

This commit adds data.line_id_files, moves the processed-line ID file
helpers there, and retargets runtime callers to the data module.

Core line-selection code now stays focused on parsing, formatting, and
range operations.
