+++ src/git_stage_batch/data/file_hunk_display.py
@@ -14,7 +14,6 @@ from ..core.diff_parser import (
     build_line_changes_from_patch_lines,
 )
 from ..core.hashing import compute_stable_hunk_hash_from_lines
-from ..core.line_selection import write_line_ids_file
 from ..core.models import (
     BinaryFileChange,
     GitlinkChange,
@@ -25,6 +24,7 @@ from ..core.models import (
     TextFileDeletionChange,
 )
 from ..core.buffer import LineBuffer
+from .line_id_files import write_line_ids_file
 from ..utils.repository_buffers import load_git_object_as_buffer
 from ..i18n import ngettext
 from ..utils.file_io import write_text_file_contents
